test<refactor>: create unique file for user stubs
