+++ test/helper/user/index.ts
@@ -1,6 +1,3 @@
-export * from './create-user.stub';
-export * from './update-user.stub';
 export * from './user-properties-generator';
-export * from './user-raw.stub';
-export * from './user-response.stub';
-
+export * from './user-stubs';
+export * from './user-matcher';
